Put user commands into its own cog
