+++ modules/core/client.py
@@ -10,6 +10,7 @@ from modules.cogs.jisho import Jisho
 from modules.cogs.daijoubu import Daijoubu
 from modules.cogs.songs import Songs
 from modules.cogs.misc import Misc
+from modules.cogs.user import User
 
 from modules.services import Service
 
@@ -44,5 +45,6 @@ class Client:
     bot.add_cog(Jisho(bot))
     bot.add_cog(Daijoubu(bot))
     bot.add_cog(Misc(bot))
+    bot.add_cog(User(bot))
 
     bot.help_command = CustomHelpCommand()
